backend/requirements.txt
alembic==1.10.2
aniso8601==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.2.4
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dotenv==1.0.0
pytz==2019.1
requests==2.28.2
six==1.12.0
SQLAlchemy==1.3.18
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==0.15.5
zipp==3.15.0
